+++ src/components/Account/AccountProfilePage.tsx
@@ -2,17 +2,19 @@ import { IDAOState, IMemberState, DAO } from "@daostack/client";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import BN = require("bn.js");
-import * as profileActions from "actions/profilesActions";
+import { getProfile, updateProfile } from "actions/profilesActions";
 import { getArc, enableWalletProvider } from "arc";
-import * as classNames from "classnames";
+import classNames from "classnames";
 import AccountImage from "components/Account/AccountImage";
 import Reputation from "components/Account/Reputation";
 import FollowButton from "components/Shared/FollowButton";
 import ThreeboxModal from "components/Shared/ThreeboxModal";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import { Field, Formik, FormikProps } from "formik";
+import Analytics from "lib/analytics";
 import { copyToClipboard, ethErrorHandler, formatTokens } from "lib/util";
-import * as queryString from "query-string";
+import { Page } from "pages";
+import { parse } from "query-string";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 import { Helmet } from "react-helmet";
@@ -36,8 +38,8 @@ interface IStateProps {
 
 interface IDispatchProps {
   showNotification: typeof showNotification;
-  getProfile: typeof profileActions.getProfile;
-  updateProfile: typeof profileActions.updateProfile;
+  getProfile: typeof getProfile;
+  updateProfile: typeof updateProfile;
 }
 
 type SubscriptionData = [IDAOState, IMemberState, BN|null, BN|null];
@@ -45,7 +47,7 @@ type IProps = IExternalProps & IStateProps & IDispatchProps & ISubscriptionProps
 
 const mapStateToProps = (state: IRootState, ownProps: IExternalProps): IExternalProps & IStateProps => {
   const accountAddress = ownProps.match.params.accountAddress ? ownProps.match.params.accountAddress.toLowerCase() : null;
-  const queryValues = queryString.parse(ownProps.location.search);
+  const queryValues = parse(ownProps.location.search);
   const daoAvatarAddress = queryValues.daoAvatarAddress as string;
 
   return {
@@ -59,8 +61,8 @@ const mapStateToProps = (state: IRootState, ownProps: IExternalProps): IExternal
 };
 
 const mapDispatchToProps = {
-  getProfile: profileActions.getProfile,
-  updateProfile: profileActions.updateProfile,
+  getProfile,
+  updateProfile,
   showNotification,
 };
 
@@ -93,6 +95,15 @@ class AccountProfilePage extends React.Component<IProps, IState> {
     const { accountAddress, getProfile } = this.props;
 
     getProfile(accountAddress);
+
+    const dao = this.props.data[0];
+
+    Analytics.track("Page View", {
+      "Page Name": Page.AccountProfile,
+      "DAO Address": dao.address,
+      "DAO Name": dao.name,
+      "Profile Address": this.props.accountAddress,
+    });
   }
 
   public copyAddress = (e: any): void => {
@@ -311,7 +322,7 @@ const SubscribedAccountProfilePage = withSubscription({
   createObservable: (props: IProps) => {
     const arc = getArc();
 
-    const queryValues = queryString.parse(props.location.search);
+    const queryValues = parse(props.location.search);
     const daoAvatarAddress = queryValues.daoAvatarAddress as string;
     const accountAddress = props.match.params.accountAddress;
     let dao: DAO;