+++ src/components/Dao/DaoSchemesPage.tsx
@@ -2,7 +2,9 @@ import { IDAOState, Scheme } from "@daostack/client";
 import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import UnknownSchemeCard from "components/Dao/UnknownSchemeCard";
+import Analytics from "lib/analytics";
 import { KNOWN_SCHEME_NAMES, PROPOSAL_SCHEME_NAMES } from "lib/util";
+import { Page } from "pages";
 import * as React from "react";
 import { BreadcrumbsItem } from "react-breadcrumbs-dynamic";
 import { RouteComponentProps } from "react-router-dom";
@@ -34,6 +36,14 @@ type IExternalProps = {
 type IProps = IExternalProps & ISubscriptionProps<Scheme[]>;
 
 class DaoSchemesPage extends React.Component<IProps, null> {
+
+  public componentDidMount() {
+    Analytics.track("Page View", {
+      "Page Name": Page.DAOSchemes,
+      "DAO Address": this.props.daoState.address,
+      "DAO Name": this.props.daoState.name,
+    });
+  }
 
   public render() {
     const { data } = this.props;