+++ src/components/Proposal/Create/SchemeForms/CreateSchemeRegistrarProposal.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { IProposalType, ISchemeState, Scheme } from "@daostack/client";
 import { enableWalletProvider, getArc } from "arc";
-import * as classNames from "classnames";
+import classNames from "classnames";
 import { ErrorMessage, Field, Form, Formik, FormikProps } from "formik";
 
 import Loading from "components/Shared/Loading";
@@ -10,8 +10,9 @@ import withSubscription, { ISubscriptionProps } from "components/Shared/withSubs
 import TagsSelector from "components/Proposal/Create/SchemeForms/TagsSelector";
 import TrainingTooltip from "components/Shared/TrainingTooltip";
 
-import * as arcActions from "actions/arcActions";
+import { createProposal } from "actions/arcActions";
 import { showNotification, NotificationStatus } from "reducers/notifications";
+import Analytics from "lib/analytics";
 import { schemeNameAndAddress, isValidUrl, GetSchemeIsActiveActions, getSchemeIsActive } from "lib/util";
 import { exportUrl, importUrlValues } from "lib/proposalUtils";
 import * as css from "../CreateProposal.scss";
@@ -25,12 +26,12 @@ interface IExternalProps {
 }
 
 interface IDispatchProps {
-  createProposal: typeof arcActions.createProposal;
+  createProposal: typeof createProposal;
   showNotification: typeof showNotification;
 }
 
 const mapDispatchToProps = {
-  createProposal: arcActions.createProposal,
+  createProposal,
   showNotification,
 };
 
@@ -83,7 +84,7 @@ class CreateSchemeRegistrarProposal extends React.Component<IProps, IState> {
       },
       title: "",
       url: "",
-      currentTab: "",
+      currentTab: "addScheme",
       tags: [],
     });
     this.state = {
@@ -91,8 +92,8 @@ class CreateSchemeRegistrarProposal extends React.Component<IProps, IState> {
       tags: this.initialFormValues.tags,
     };
   }
-  
-  
+
+
   public async handleSubmit(values: IFormValues, { setSubmitting }: any ):  Promise<void> {
     if (!await enableWalletProvider({ showNotification: this.props.showNotification })) { return; }
 
@@ -133,6 +134,14 @@ class CreateSchemeRegistrarProposal extends React.Component<IProps, IState> {
     };
     setSubmitting(false);
     await this.props.createProposal(proposalValues);
+
+    Analytics.track("Submit Proposal", {
+      "DAO Address": this.props.daoAvatarAddress,
+      "Proposal Title": values.title,
+      "Scheme Address": this.props.scheme.address,
+      "Scheme Name": this.props.scheme.name,
+    });
+
     this.props.handleClose();
   }
 
@@ -152,7 +161,7 @@ class CreateSchemeRegistrarProposal extends React.Component<IProps, IState> {
     exportUrl(values);
     this.props.showNotification(NotificationStatus.Success, "Exportable url is now in clipboard :)");
   }
-  
+
   public render(): RenderOutput {
     // "schemes" are the schemes registered in this DAO
     const schemes = this.props.data;