+++ src/components/Proposal/Create/SchemeForms/CreateUnknownGenericSchemeProposal.tsx
@@ -1,7 +1,8 @@
 import { ISchemeState } from "@daostack/client";
-import * as arcActions from "actions/arcActions";
+import { createProposal } from "actions/arcActions";
 import { enableWalletProvider } from "arc";
 import { ErrorMessage, Field, Form, Formik, FormikProps } from "formik";
+import Analytics from "lib/analytics";
 import * as React from "react";
 import { connect } from "react-redux";
 import { showNotification, NotificationStatus } from "reducers/notifications";
@@ -19,7 +20,7 @@ interface IExternalProps {
 }
 
 interface IDispatchProps {
-  createProposal: typeof arcActions.createProposal;
+  createProposal: typeof createProposal;
   showNotification: typeof showNotification;
 }
 
@@ -30,7 +31,7 @@ interface IStateProps {
 type IProps = IExternalProps & IDispatchProps;
 
 const mapDispatchToProps = {
-  createProposal: arcActions.createProposal,
+  createProposal,
   showNotification,
 };
 
@@ -59,7 +60,7 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
       value: 0,
       tags: [],
     });
-    this.state = { 
+    this.state = {
       tags: this.initialFormValues.tags,
     };
   }
@@ -75,6 +76,14 @@ class CreateGenericScheme extends React.Component<IProps, IStateProps> {
 
     setSubmitting(false);
     await this.props.createProposal(proposalValues);
+
+    Analytics.track("Submit Proposal", {
+      "DAO Address": this.props.daoAvatarAddress,
+      "Proposal Title": values.title,
+      "Scheme Address": this.props.scheme.address,
+      "Scheme Name": this.props.scheme.name,
+    });
+
     this.props.handleClose();
   }