+++ src/pages.ts
@@ -1,0 +1,12 @@
+export enum Page {
+  AccountProfile = "Account Profile",
+  AllDAOs = "All Daos",
+  CreateProposal = "Create Proposal",
+  DAOHistory = "DAO History",
+  DAOMembers = "DAO Members",
+  DAOSchemes = "DAO Schemes",
+  ProposalDetails = "Proposal Details",
+  Redemptions = "Redemptions",
+  RedemptionsMenu = "Redemptions Menu",
+  SchemeProposals = "Scheme Proposals",
+}